second commit test

+++ recette/src/main/java/com/example/recette/dao/EtapeDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EtapeDao  extends JpaRepository<Etape,Integer> {
+public interface EtapeDao extends JpaRepository<Etape, Long> {
     int deleteByRef(String ref);
-
 }
+
